% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adorn.R
\name{cat_adorn_perc}
\alias{cat_adorn_perc}
\title{Adorning cats}
\usage{
cat_adorn_perc(x, sep = " ", parentheses = TRUE)
}
\arguments{
\item{x}{a categorical variable (i.e., a factor)}

\item{sep}{a string that will be placed between the current
levels of `x` and the adorning proportions.}

\item{parentheses}{`TRUE` or `FALSE`. If `TRUE`, parentheses
are placed around the percentage values. If `FALSE`, then
nothing is placed around the percentage values.}
}
\value{
a factor with new levels
}
\description{
For graphs or tables, categorical variables are often accompanied
  by sample sizes and corresponding proportions for each category.
  For example, group A had 10 participants, accounting for 50%
  of the A versus B study. These functions add descriptive
  counts and proportions to existing categorical variables
  so that presenting these data is easier.
}
\examples{

x <- factor(letters[c(1,1,1,1, 2,2,2,2)])
cat_adorn_perc(x, parentheses = FALSE)
cat_adorn_perc(cat_adorn_nobs(x))


}
