% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herd.R
\name{cat_gather}
\alias{cat_gather}
\title{Herd Cats}
\usage{
cat_gather(data, factor_levels)
}
\arguments{
\item{data}{Data with one-hot encoded variables that have names formatted
as <NAME>_<CATEGORY>. For example, `x_a` and `x_b` should be column
names for a one hot encoded variable `x` with levels `'a'` and `'b'`.}

\item{factor_levels}{a list with names and values corresponding to
factor names and factor levels, respectively. For example,
`list(x = c('a','b'))` shows the expected input to gather a
categorical variable named `x` with levels `'a'` and `'b'`}
}
\description{
many modeling functions expect matrix input with
  factor levels one-hot encoded.

  - [cat_spread] will one-hot encode any factor or character variable
  in `data` and return a one-hot encoded `tibble`.

  - `cat_gather` will apply the inverse operation and convert one-hot
  encoded columns back into factors.
}
