% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herd.R
\name{cat_spread}
\alias{cat_spread}
\title{Herd Cats}
\usage{
cat_spread(
  data,
  ...,
  sparsify_nas = FALSE,
  na_cols = FALSE,
  drop_cols = TRUE,
  drop_unused_levels = FALSE
)
}
\arguments{
\item{data}{Data with categorical variables (i.e., factors)
that need to be spread into one-hot format.}

\item{...}{Column names to spread into one-hot format. This can also be
a tidyselect specification (e.g., `starts_with()`, `contains()`, etc.).
If unspecified, all character and factor variables in `data` will be
spread into one-hot format.}

\item{sparsify_nas}{(logical) Should NAs be converted to 0s?.
Default is `FALSE`.}

\item{na_cols}{(logical) Should columns be generated to indicate
the presence of NAs? This argument only applies to factor columns
with at least one NA. Default is `FALSE`.}

\item{drop_cols}{(logical) Should the resulting data frame exclude
the original columns that were spread into one-hot format? Default
is `TRUE`.}

\item{drop_unused_levels}{(logical) Should columns of all 0s be
dropped? This occurs if `data` has unused factor levels. Default
is `FALSE`.}
}
\value{
a [tibble][tibble::tibble-package] with categorical variables
  herded as you like.
}
\description{
many modeling functions expect matrix input with
  factor levels one-hot encoded.

  - `cat_spread` will one-hot encode any factor or character variable
  in `data` and return a one-hot encoded `tibble`.

  - [cat_gather] will apply the inverse operation and convert one-hot
  encoded columns back into factors.
}
\examples{

df <- data.frame(x = rep(letters[1:2], 50), y = 1:100)

one_hot_df <- cat_spread(df)
print(one_hot_df)

cat_gather(one_hot_df, factor_levels = list(x=c('a','b')))


}
